+++ src/components/Category/CategoryHeader.tsx
@@ -3,7 +3,7 @@ import { Stack, Typography } from "@mui/material";
 import { skipToken } from "@reduxjs/toolkit/dist/query/react";
 import { useRouter } from "next/router";
 import useAddDefaultSortType from "../../hooks/sort";
-import { useGetSingleCategoryQuery } from "../../services/productsService";
+import { useGetSingleCategoryQuery } from "../../services/products";
 import { removeColor, removeSize, setColor, setSize } from "../../store/slices/filterSlice";
 import Filter from "./Filters/Filter";
